planetrush_flask/requirements.txt
blinker==1.8.2
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.7
colorama==0.4.6
filelock==3.15.4
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
fsspec==2024.6.1
greenlet==3.0.3
idna==3.4
itsdangerous==2.2.0
Jinja2==3.1.4
joblib==1.4.2
MarkupSafe==2.1.5
mpmath==1.3.0
networkx==3.3
numpy==1.26.4
pillow==10.4.0
PyMySQL==1.1.1
python-dotenv==1.0.1
python-mecab-ko==1.3.7
python-mecab-ko-dic==2.1.1.post2
requests==2.28.1
scikit-learn==1.5.1
scipy==1.14.0
SQLAlchemy==2.0.32
sympy==1.13.2
threadpoolctl==3.5.0
torch==2.0.1+cu117
torchaudio==2.0.2+cu117
torchvision==0.15.2+cu117
typing_extensions==4.12.2
urllib3==1.26.13
Werkzeug==3.0.3
